fix(net): 授时自纠/int64、g_ip 锁、清表顺延、http 非阻塞等 8 处隐患（d143）
- 授时：upd_last 落在未来时差值恒负，旧写法会永久静默跳过自动检查且不自纠，
  改为负差值按"到期"做完整检查（随后落盘的真实时间即修正回来）
- 2038：天数 * 86400 全程 64 位，config 的 updLast 一并改 long long
  （只改计算不改存储，2038 后仍会被 (int) 截断成负值）
- g_ip 读写加独立锁（不复用 ctl 锁：那对函数靠全局 g_ctl_held 记"本线程
  是否持锁"，别的线程借用会让 watch 的 unlock 失效）；撕裂读到半截 IP 会让
  扫描按错的 /24 整轮白扫
- 手动扫描清表：正在扫时改为挂起到下一轮开始前，否则清完立刻被在跑的旧轮填回
- known-hosts 追加去重：config 按原字符串去重，"x.7" 与 "x.07" 解析出同一
  主机号，重复追加会写穿 hosts[254]
- Device.download 按协议取值（scan/discovery 两处），不再恒 false
- SO_NBIO 返回值判定（scan/net/discovery 三处）：置位失败即放弃本次操作，
  不让阻塞 fd 卡死主线程或扫描 worker
- http：删掉 worker fd 上语义冲突的 SO_RCVTIMEO=3s，worker 与监听 fd 统一置
  非阻塞，"3s 头截止 / 30s 收体空闲"才真正由 conn_read 自己数

TAG d143，真机验证：授时自纠（9989477937 → 1789729072）、三角清表顺延、
并发 chunked 上传 200、断网退场与恢复均正常

+++ src/core/config.c
@@ -106,7 +106,7 @@ void config_init(void)
             if (json_get_int(buf, "paneSwap", &v)) g_cfg.pane_swap = (int)v;
             if (json_get_int(buf, "lang", &v)) g_cfg.lang = (int)v;
             if (json_get_int(buf, "updateAuto", &v)) g_cfg.upd_auto = (int)v;
-            if (json_get_int(buf, "updateLast", &v)) g_cfg.upd_last = (int)v;
+            if (json_get_int(buf, "updateLast", &v)) g_cfg.upd_last = v;
             json_get_str(buf, "saveDir", g_cfg.save_dir, sizeof g_cfg.save_dir);
             {   /* knownIps: "ip,ip,..."（逗号分隔，最新在前） */
                 char k[512];
@@ -195,7 +195,7 @@ static void config_write_locked(void)
                    "  \"paneSwap\": %d,\n"
                    "  \"lang\": %d,\n"
                    "  \"updateAuto\": %d,\n"
-                   "  \"updateLast\": %d,\n"
+                   "  \"updateLast\": %lld,\n"
                    "  \"saveDir\": \"%s\",\n"
                    "  \"knownIps\": \"%s\"\n"
                    "}\n",
@@ -310,7 +310,7 @@ void config_set_upd_auto(int v)
     cfg_unlock();
 }
 
-void config_set_upd_last(int t)
+void config_set_upd_last(long long t)
 {
     cfg_lock();
     g_cfg.upd_last = t < 0 ? 0 : t;
